split objects into separate files
