+++ constants.py
@@ -11,7 +11,7 @@ EXTRA_HEALTH = 10
 BLOCK = 32
 BOMBTICKER = 10
 BOMBTIMEOUT = 2000
-BULLET_SPEED = 4
+BULLET_SPEED = 14
 BULLET_TIMER = 2
 CAMERA_SPEED = 0.1
 FLAME_RATE = 33
@@ -42,7 +42,7 @@ PLAYER_MOVEMENT_SPEED = 4
 PLAYER_SCALING = 1
 RECT_HEIGHT:int = BLOCK
 RECT_WIDTH:int = BLOCK
-SCREEN_HEIGHT = 600
+SCREEN_HEIGHT = 800
 SCREEN_TITLE = "bdude"
 SCREEN_WIDTH = 800
 SPRITE_SCALING = 1
