+++ objects/blocks.py
@@ -1,0 +1,29 @@
+from dataclasses import dataclass, field
+import asyncio
+from loguru import logger
+from pygame.math import Vector2 as Vec2d
+from pygame.sprite import Sprite
+# from pymunk import Vec2d
+import json
+import math
+import pygame
+import random
+import time
+from utils import gen_randid
+from constants import PLAYER_MOVEMENT_SPEED, PARTICLE_COUNT, PARTICLE_RADIUS, PARTICLE_SPEED_RANGE, PARTICLE_MIN_SPEED, PARTICLE_FADE_RATE, PARTICLE_GRAVITY, FLAME_SPEED, FLAME_TIME, FLAME_RATE, BOMBTICKER, BULLET_TIMER, FLAMEX, FLAMEY
+
+class Upgrade(Sprite):
+	def __init__(self, upgradetype, image, position, scale, timer=1000):
+		super().__init__()
+		self.image = pygame.image.load(image)
+		self.rect = self.image.get_rect()
+		self.upgradetype = upgradetype
+		self.position = position
+		self.rect.topleft = self.position
+		self.timer = timer
+
+	def update(self):
+		if self.timer <= 0:
+			self.kill()
+		else:
+			self.timer -= 1
